Fix for build break caused by annotation changes to interface descriptions Fix for some compiler warnings under OSX

+++ AllJoynFramework/AllJoynFramework/AJNBus.h
@@ -21,7 +21,7 @@
  * org.freedesktop.DBus.RequestName input params (see DBus spec)
  */
 // @{
-typedef NSInteger AJNBusNameFlag;
+typedef uint32_t AJNBusNameFlag;
 
 /**< RequestName input flag: Allow others to take ownership of this name */
 extern const AJNBusNameFlag kAJNBusNameFlagAllowReplacement;
